src/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-ai-projects>=2.0.0b3
azure-identity
azure-keyvault-secrets
